+++ Changelog.md
@@ -27,8 +27,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add PARANOID relocation table overflow check to GC sweep phase.
 - Option VERBOSE\_COMPRESSED\_CONS for diagnostic printing a 'C' for
   each compressed cons.
-- Breakpoints on user-defined procedures. (Symbol list "\*b\*".)
-- Debugger keeps its current return value in symbol "\*r\*".
+- Breakpoints in debugger.
+- Improved debugger REPL return value handling.
 
 #### Changed